[FIX+IMP] currency_rate_update: Several things:
* Update to latest OCA conventions

  Update README
  Code cleanup: remove @api.one, better error messages, remove call to MOD_NAME which was not defined any more
  Use account.menu_config_multi_currency as parent menu
* Remove res.company view
* Crash when you add currencies to "Currencies to update with this service"
* should now work in multi-company
