+++ currency_rate_update/README.rst
@@ -1,10 +1,12 @@
 .. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-    :alt: License
+   :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+   :alt: License: AGPL-3
 
+====================
 Currency Rate Update
 ====================
 
-Import exchange rates from the Internet.
+Download exchange rates automatically from the Internet.
 
 The module is able to use the following sources:
 
@@ -38,13 +40,11 @@ The module is able to use the following sources:
 Configuration
 =============
 
-The update can be set under the company form.
-You can set for each services which currency you want to update.
-The logs of the update are visible under the service note.
-You can active or deactivate the update.
-The module uses internal ir-cron feature from Odoo, so the job is
-launched once the server starts if the 'first execute date' is before
-the current day.
+To configure the module, go to *Accounting > Configuration > Multi-currencies > Rate Auto-download* and create one or several services to download rates from the Internet.
+
+Then, go to the page *Accounting > Configuration > Settings* and, in the section *Multi Currencies*, make sure that the option *Automatic Currency Rates Download* is enabled.
+
+In developper mode, in the menu *Settings > Technical > Scheduled Actions*, make sure that the action *Currency Rate Update* is active. If you want to run it immediately, use the button *Run Manually*.
 
 Usage
 =====
@@ -72,14 +72,12 @@ Roadmap:
 * Updated daily from Citibank N.A., source in EUR. Information may be delayed.
   This is parsed from an HTML page, so it may be broken at anytime.
 
-
 Bug Tracker
 ===========
 
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/account-financial-tools/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed feedback.
-
 
 Credits
 =======
@@ -108,13 +106,12 @@ Contributors
 * Daniel Dico <ddico@oerp.ca> (BOC)
 * Dmytro Katyukha <firemage.dima@gmail.com>
 
-
 Maintainer
 ----------
 
-.. image:: http://odoo-community.org/logo.png
+.. image:: https://odoo-community.org/logo.png
    :alt: Odoo Community Association
-   :target: http://odoo-community.org
+   :target: https://odoo-community.org
 
 This module is maintained by the OCA.
 
@@ -122,4 +119,4 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-To contribute to this module, please visit http://odoo-community.org.
+To contribute to this module, please visit https://odoo-community.org.
