+++ currency_rate_update/__init__.py
@@ -1,2 +1,4 @@
-from . import model
+# -*- coding: utf-8 -*-
+
+from . import models
 from .services.currency_getter_interface import CurrencyGetterInterface
