+++ currency_rate_update/__manifest__.py
@@ -13,13 +13,11 @@
         "account",  # Added to ensure account security groups are present
     ],
     "data": [
-        "view/service_cron_data.xml",
-        "view/currency_rate_update.xml",
-        "view/company_view.xml",
+        "data/cron.xml",
+        "views/currency_rate_update.xml",
+        "views/account_config_settings.xml",
         "security/rule.xml",
         "security/ir.model.access.csv",
     ],
-    "images": [],
-    "demo": [],
     'installable': True
 }
