--- currency_rate_update/model/__init__.py
@@ -1,2 +1,0 @@
-from . import currency_rate_update
-from . import company
