+++ currency_rate_update/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from . import currency_rate_update
+from . import company
+from . import account_config_settings
