+++ currency_rate_update/models/account_config_settings.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+# © 2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from odoo import models, fields
+
+
+class AccountConfigSettings(models.TransientModel):
+    _inherit = 'account.config.settings'
+
+    auto_currency_up = fields.Boolean(
+        related='company_id.auto_currency_up')
