+++ currency_rate_update/services/currency_getter_interface.py
@@ -5,8 +5,8 @@
 import logging
 
 from datetime import datetime
-from odoo import fields
-from odoo.exceptions import except_orm
+from odoo import fields, _
+from odoo.exceptions import UserError
 
 _logger = logging.getLogger(__name__)
 
@@ -143,15 +143,11 @@ class CurrencyGetterInterface(object):
             objfile.close()
             return rawfile
         except ImportError:
-            raise except_orm(
-                'Error !',
-                self.MOD_NAME + 'Unable to import urllib !'
-            )
+            raise UserError(
+                _('Unable to import urllib.'))
         except IOError:
-            raise except_orm(
-                'Error !',
-                self.MOD_NAME + 'Web Service does not exist !'
-            )
+            raise UserError(
+                _('Web Service does not exist (%s)!') % url)
 
     def check_rate_date(self, rate_date, max_delta_days):
         """Check date constrains. rate_date must be of datetime type"""
@@ -160,9 +156,9 @@ class CurrencyGetterInterface(object):
             raise Exception(
                 'The rate timestamp %s is %d days away from today, '
                 'which is over the limit (%d days). '
-                'Rate not updated in OpenERP.' % (rate_date,
-                                                  days_delta,
-                                                  max_delta_days)
+                'Rate not updated in Odoo.' % (rate_date,
+                                               days_delta,
+                                               max_delta_days)
             )
 
         # We always have a warning when rate_date != today
