--- currency_rate_update/view/currency_rate_update.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-    <data>
-        <record model="ir.ui.view" id="currency_rate_update_tree">
-            <field name="name">Update Rates service</field>
-            <field name="model">currency.rate.update.service</field>
-            <field name="arch" type="xml">
-                <tree string="Currency update services">
-                    <field name="service"/>
-                    <field name="company_id" groups="base.group_multi_company"
-                        invisible="not context.get('currency_rate_update_main_view')"/>
-                    <field name="currency_to_update"/>
-                    <field name="interval_type"/>
-                    <field name="interval_number"/>
-                    <field name="next_run"/>
-                </tree>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="currency_rate_update_form">
-            <field name="name">Update Rates</field>
-            <field name="model">currency.rate.update.service</field>
-            <field name="arch" type="xml">
-                <form string="Currency update services">
-                    <group name="main">
-                        <group name="left">
-                            <field name="service"/>
-                            <field name="company_id" groups="base.group_multi_company"
-                                invisible="not context.get('currency_rate_update_main_view')"
-                                required="context.get('currency_rate_update_main_view')"/>
-                            <field name="max_delta_days"/>
-                            <button name="refresh_currency" type="object"
-                                    string="Update now!"/>
-                        </group>
-                        <group name="right">
-                            <field name="interval_type"/>
-                            <field name="interval_number"/>
-                            <field name="next_run"/>
-                        </group>
-                    </group>
-                    <group string="Currencies to update with this service">
-                        <field name="currency_list" invisible="1"/>
-                        <field name="currency_to_update" nolabel="1" domain="[('id','in', currency_list[0][2])]"/>
-                    </group>
-                    <group string="Logs" name="logs">
-                        <field name="note" nolabel="1"/>
-                    </group>
-                </form>
-            </field>
-        </record>
-
-        <record id="currency_rate_update_action" model="ir.actions.act_window">
-            <field name="name">Currency Rate Update</field>
-            <field name="res_model">currency.rate.update.service</field>
-            <field name="view_mode">tree,form</field>
-            <field name="context">{'currency_rate_update_main_view': True}</field>
-        </record>
-
-        <menuitem id="currency_rate_update_menu" action="currency_rate_update_action"
-            parent="account.menu_finance_configuration" sequence="21"/>
-
-    </data>
-</odoo>
