+++ currency_rate_update/views/account_config_settings.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  © 2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+
+<record id="view_account_config_settings" model="ir.ui.view">
+    <field name="name">currency_rate_update.account_config_settings.form</field>
+    <field name="model">account.config.settings</field>
+    <field name="inherit_id" ref="account.view_account_config_settings"/>
+    <field name="arch" type="xml">
+        <group name="multi_currency" position="after">
+            <group name="currency_rate_update">
+                <label for="id" string="Auto-download Rates"/>
+                <div name="currency_rate_update">
+                    <field name="auto_currency_up" class="oe_inline"/>
+                    <label for="auto_currency_up"/>
+                </div>
+            </group>
+        </group>
+    </field>
+</record>
+
+</odoo>
